refactor(studio.giselles.ai): remove obsolete handle_event function
